dont add an array axis when there is only 1 flag
Also take this opportunity to split out the _unpack_to_array method
to a core method and a user facing method.
Now it's more straightforward for _unpack_to_dict to call the
core method and adjust the output as needed.

Also some new tests to go along with the changes.